chatbot-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.4.2
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
numpy>=1.24.0
pandas>=2.0.0
